Parampampam

+++ src/Swtg1337Digits.java
@@ -4,9 +4,9 @@ public class Swtg1337Digits{
     public static void main(String args[]){
        Scanner reader = new Scanner(System.in);
        System.out.println("Введите число х:");
-       int x =reader.nextInt();
+       double x =reader.nextDouble();
        System.out.println("Введите число y:");
-       int y =reader.nextInt();
+       double y =reader.nextDouble();
        if (x==y) {System.out.println("Числа равны");}
        else {System.out.println("Числа не равны.");};
        System.out.println("#swtg1337");
